+++ hook_process.c
@@ -368,7 +368,7 @@ HOOKDEF(NTSTATUS, WINAPI, NtUnmapViewOfSection,
     _In_      HANDLE ProcessHandle,
     _In_opt_  PVOID BaseAddress
 ) {
-    unsigned int map_size = 0; MEMORY_BASIC_INFORMATION mbi;
+    SIZE_T map_size = 0; MEMORY_BASIC_INFORMATION mbi;
     if(VirtualQueryEx(ProcessHandle, BaseAddress, &mbi,
             sizeof(mbi)) == sizeof(mbi)) {
         map_size = mbi.RegionSize;
